+++ CHANGELOG.md
@@ -8,6 +8,10 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+[PR-11](https://github.com/itk-dev/itqr/pull/11)
+    - Changed url and qr entity models
+[PR-9](https://github.com/itk-dev/itqr/pull/9)
+    - Added bulk change url
 [PR-6](https://github.com/itk-dev/itqr/pull/6)
     - Setup some easyadmin stuff
 [PR-4](https://github.com/itk-dev/itqr/pull/4)